Limit server response propagation to a set interval (#1299)
When we have large clusters that produce high rates of change, there's
no point in sending every broadcasted update over the network to any
subscribing client. By making sure that we only ever propagate change
updates at a set interval we can avoid unnecessary work and reduce
pressure on the system as a whole.

The 100ms interval is purely arbitrary and could be set lower if we
wish.