Enable continuous memory profiling (#1306)
Adds `jemalloc` feature and enables it by default, which uses the
jemalloc allocator and allows us to expose `/debug/pprof/allocs` on the
admin endpoint to get `inuse_space` memory profiles, that can be
automatically scraped into grafana.

Had to move some things around to not collide with the `mimalloc` and
`heap-stats` features. I don't know the use-cases and/or status of those
features though. On MSVC the `jemalloc` feature should be a no-op since
it appears to not be supported.