Expose jemalloc statistics as prometheus metrics, remove `quilkin` prefix from select metrics (#1315)
Besides adding jemalloc statistics as metrics, this removes the blanket
`quilkin` prefix from the metric registry, and instead adds the prefix
to the metrics individually. This will ensure that other metrics which
are not necessarily specific to quilkin functionality like `jemalloc_`,
`http_` and `xds_` can be exported without the quilkin prefix. I added
the `xds_` prefix to the few xds crate metrics that did not have it.